feature: Setting SSR for blog posts
I was doing page rendering without SSR which is ok for non static pages, but not for blogs and content pages.
It also demonstrated some Astro routing features that I wanted to share with the viewers.
Now we have and understand both methods, which give us the flexibility on the hybrid output in Astro.